refactor(server): #125 — extract exec handlers into routes/exec.rs
